net: shell: Add dns command
Move "dns" command to a separate file.
