+++ subsys/net/lib/shell/CMakeLists.txt
@@ -10,6 +10,7 @@ zephyr_library_sources(allocs.c)
 zephyr_library_sources(arp.c)
 zephyr_library_sources(capture.c)
 zephyr_library_sources(conn.c)
+zephyr_library_sources(dns.c)
 
 zephyr_library_include_directories(
   ${ZEPHYR_BASE}/kernel/include
